backend/requirements.txt
fastapi
uvicorn
pypetkitapi
python-dotenv
sqlmodel
aiohttp
